+++ java/src/test/java/io/github/linkfoundation/linksnotation/IndentedIdNestedValuesTest.java
@@ -1,0 +1,52 @@
+package io.github.linkfoundation.linksnotation;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+class IndentedIdNestedValuesTest {
+  @Test
+  void indentedIdRetainsNestedValues() throws ParseException {
+    String[][] cases = {
+      {
+        "outer:\n  inner:\n    value1\n    value2\n  value3",
+        "(outer: (inner: value1 value2) value3)"
+      },
+      {
+        "statement:\n"
+            + "  subject:\n"
+            + "    I\n"
+            + "  verb:\n"
+            + "    love\n"
+            + "  object:\n"
+            + "    you:\n"
+            + "      very\n"
+            + "      much",
+        "(statement: (subject: I) (verb: love) (object: (you: very much)))"
+      },
+      {
+        "document:\n"
+            + "  (metadata: title author date)\n"
+            + "  content:\n"
+            + "    paragraph1\n"
+            + "    (paragraph2: text (with: nested structure))\n"
+            + "    paragraph3",
+        "(document: (metadata: title author date) (content: paragraph1 (paragraph2: text (with:"
+            + " nested structure)) paragraph3))"
+      },
+      {
+        "image:\n    file: .gitpod.Dockerfile\n    context: ./docker-content",
+        "(image: (file: .gitpod.Dockerfile) (context: ./docker-content))"
+      },
+      {
+        "level1:\n  level2:\n    level3a\n    level3b\n  level2b",
+        "(level1: (level2: level3a level3b) level2b)"
+      },
+      {"root:\n  child1\n  child2\n    grandchild", "(root: child1 (child2 grandchild))"}
+    };
+    Parser parser = new Parser();
+    for (String[] testCase : cases) {
+      assertEquals(testCase[1], Link.formatLinks(parser.parse(testCase[0])), testCase[0]);
+    }
+  }
+}